refactor(areacode-map): split map page by responsibilities
